feat(react): provide docgen via a worker-target module
The React renderer contributes a DocgenProviderDescriptor pointing at a
worker-target module that runs react-component-meta extraction inside core's
docgen worker. The renderer owns no threading code. Extraction builds its own
ComponentMetaManager in the worker, so the manifest generator's shared singleton
is now scoped to the manifest path only.
