+++ code/renderers/react/src/docgen/preset.ts
@@ -1,58 +1,24 @@
-import { STORY_FILE_TEST_REGEXP, getStoryImportPathFromEntry } from 'storybook/internal/common';
-import type { DocgenProviderPreset } from 'storybook/internal/types';
+import { fileURLToPath } from 'node:url';
 
-import { getSharedComponentMetaManager } from '../componentManifest/componentMetaManagerSingleton.ts';
-import type { TypescriptOptions } from '../componentManifest/getComponentImports.ts';
-import { buildDocgenPayload } from './buildDocgen.ts';
+import type { DocgenProviderDescriptor } from 'storybook/internal/types';
 
 /**
- * React renderer docgen provider — phase 3: real RCM-backed extraction.
+ * React renderer docgen provider.
  *
- * Receives the authoritative index `entry` from the docgen service. Bails to `nextDocgen` when
- * the entry does not resolve to a CSF story file and when TypeScript isn't available. Otherwise
- * delegates to {@link buildDocgenPayload} which runs RCM against the file and returns a complete
- * {@link DocgenPayload}, or falls through to `nextDocgen` when nothing extractable is found.
+ * Contributes a {@link DocgenProviderDescriptor} pointing at {@link ./docgen-worker.ts}, which
+ * core's docgen worker imports and runs off the main thread. The renderer owns no threading code —
+ * it only resolves the worker module path; nothing else needs to cross the worker boundary, since
+ * `react-component-meta` extraction reads nothing from the user's `typescript` options (see
+ * {@link ./docgen-worker.ts}).
  *
- * When extraction succeeds, the payload is merged with downstream via the documented
- * `{ ...downstream, ...ours }` spread idiom so any fields a future provider sets and we don't
- * know about survive intact.
+ * The descriptor is appended to the accumulated array so addon providers can stack on top, mirroring
+ * the previous middleware chain — just composed inside the worker instead of in this process.
  */
-export const experimental_docgenProvider: DocgenProviderPreset = async (nextDocgen, options) => {
-  // Resolve the typescript options preset once at chain-build time so the provider closure does
-  // not re-read it on every call.
-  const typescriptOptionsPromise = (options.presets?.apply<Partial<TypescriptOptions>>(
-    'typescript',
-    {}
-  ) ?? Promise.resolve({})) as Promise<Partial<TypescriptOptions>>;
-
-  // Pre-boot the component meta manager in the background, off the critical path.
-  // `import('typescript')` loads a multi-MB module; deferring it to the first docgen request adds
-  // several seconds of latency to that request. Kicking it off here (fire-and-forget) lets it warm
-  // up in parallel with the rest of server startup, so `await getSharedComponentMetaManager()`
-  // below usually resolves instantly. The singleton memoizes and swallows its own errors, so this
-  // is safe to leave unawaited.
-  void getSharedComponentMetaManager();
-
-  return async (input) => {
-    const storyImportPath = getStoryImportPathFromEntry(input.entry);
-    if (!storyImportPath || !STORY_FILE_TEST_REGEXP.test(storyImportPath)) {
-      return nextDocgen(input);
-    }
-
-    const componentMetaManager = await getSharedComponentMetaManager();
-    if (!componentMetaManager) {
-      return nextDocgen(input);
-    }
-
-    const ours = await buildDocgenPayload(input, {
-      componentMetaManager,
-      typescriptOptions: await typescriptOptionsPromise,
-    });
-    if (!ours) {
-      return nextDocgen(input);
-    }
-
-    const downstream = await nextDocgen(input);
-    return { ...downstream, ...ours };
-  };
-};
+export const experimental_docgenProvider = async (
+  existing: DocgenProviderDescriptor[] = []
+): Promise<DocgenProviderDescriptor[]> => [
+  ...existing,
+  {
+    moduleSpecifier: fileURLToPath(import.meta.resolve('@storybook/react/internal/docgen-worker')),
+  },
+];
